% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTwo.r
\name{appTwo}
\alias{appTwo}
\title{example Shiny app}
\usage{
appTwo(results)
}
\arguments{
\item{results}{a results object}
}
\description{
A simple demonstration application to test the elements of the library
}
