% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableDataSummary.r
\name{availableDataSummary}
\alias{availableDataSummary}
\title{summary of the available data and selections}
\usage{
availableDataSummary(id, input, output, session, ui = T, cx = NULL)
}
\arguments{
\item{cx}{}
}
\description{
summary of the available data and selections
}
