% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPcaViewConnectivity.r
\name{checkPcaViewConnectivity}
\alias{checkPcaViewConnectivity}
\title{checks to see if PCA views are compariable to triangulate a single direction}
\usage{
checkPcaViewConnectivity(w)
}
\arguments{
\item{w}{table of input projections of the form simulated by \link{makeW}}
}
\value{
a list with two slots, selected and missing: selected gives a vector of the dimensions in which the vector is assumed
   to lie. If a single direction is described, missing has length zero - list(). If more than one vector is implied, two vectors -
   group1 and group2 are returned given the dimensions of the subspaces what have to be linked to define a single direction.
}
\description{
Consider a six dimensional space, but we only have views for directions 1, 2 and 3, and 5 and 6.
In this case we (i) only have information about dimensions 1, 2, 3, 5 and 6 and (ii) have a suggested direction for TWO vectors,
one in dimensions 1-3 and a second in 4-6. This function (a) takes a collection of projected directions, returns the dimensions
the hidden vector lies in [i.e. those that define its hyperplane or subspace] and (c) if multiple vectors in different subspaces are
implied, as list of two sets of accross which projections should be returned to define a single vector.
}
\examples{

w=makeW(1:6)
checkPcaViewConnectivity(w)

w[,1:2]
checkPcaViewConnectivity(w[,1:2])

w[,c(1:2,15)]
checkPcaViewConnectivity(w[,c(1:2,15)])
}
