% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTwoWayVennDiagram.r
\name{computeTwoWayVennDiagram}
\alias{computeTwoWayVennDiagram}
\title{computes a tabular two-way Venn diagram from a list of contrast objects}
\usage{
computeTwoWayVennDiagram(Q)
}
\arguments{
\item{Q}{a list of goViewerInput objects all of type 'split'}
}
\value{
a matrix of cross tabulations of frequency counts.
}
\description{
computes a tabular two-way Venn diagram from a list of contrast objects
}
\examples{
data(P2)
Q2=P2[1:3]

demo=function(x,a=500,b=1000){
  x$type="split"
  x$data=0*x$data
  x$data[a:b]=x$data[a:b]+1
  x
}

Q2=list(demo(P2[[1]],100,500),
        demo(P2[[2]],100,1000),
        demo(P2[[3]],5000,6000))


computeTwoWayVennDiagram(Q2)
computeTwoWayVennDiagram(Q2[1]) # also works for one selection
computeTwoWayVennDiagram(list()) # but not an empty list

Q2[[1]]$type="ordered"
computeTwoWayVennDiagram(Q2)
}
