% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVennDiagram.r
\name{computeVennDiagram}
\alias{computeVennDiagram}
\title{computes a full Venn diagram}
\usage{
computeVennDiagram(Q)
}
\arguments{
\item{Q}{a named list of sets. If no names are given, default names are added}
}
\value{
a matrix with as many columns as logical combinations and as many rows as there are input sets. The column names give the sums
  the corresponding combinations and the row names, the names of the sets.
}
\description{
Note , to test, draw out a diagram first, populate with some random [unique] symbols, and then
note down the original sets.
}
\examples{

computeVennDiagram(list(A=c(4,5,6,7),
                        B=c(5,6,7,8,9,10),
                        C=c(7,10)))

computeVennDiagram(list(c(4,1,2,3,5,6,7),
                        c(1,2,3,5,6,7,9),
                        c(5,6,7,9,8)))
}
