% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseAddActions.r
\name{condenseAddActions}
\alias{condenseAddActions}
\title{integrates over a list of set add and delete actions}
\usage{
condenseAddActions(actions, last = F)
}
\arguments{
\item{actions}{a list of lists; each sub list containing a action slot - "add" or "remove"
and a "names" slot containing the items to be added or removed}

\item{last}{if true, a string summarising the last action is returned}
}
\value{
a vector the the remaining items
}
\description{
Designed to be used for point brushing, this allows for both an add
and a delete mode. Note that the requests have to be processed in
order. Changing the order can change the results.
}
\details{
NOTE THAT THIS FUNCTION IS DESIGNED TO PARSE AND READ "events" NOT WRITE TO IT.
UYOU CAN DO THIS DIRECTLY IN ANY REACTIVE BLOCK.
}
\examples{

actions=list(
  list(action="add",
       names=c("A","B")),
  list(action="add",
       names=c("C")),
  list(action="remove",
       names=c("B","D"))
)

actions
condenseAddActions(actions)
condenseAddActions(actions,last=T)
}
