% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastViewer.r
\name{contrastViewer}
\alias{contrastViewer}
\title{top level contrast viewer code}
\usage{
contrastViewer(id, input, output, session = NULL, ui = T, cx = NULL)
}
\arguments{
\item{cx}{}
}
\value{

}
\description{
top level contrast viewer code
}
