% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countUniverse.r
\name{countUniverse}
\alias{countUniverse}
\title{counts the total number of observed and selected entries}
\usage{
countUniverse(Q)
}
\arguments{
\item{Q}{a list of goViewer 'split' objects}
}
\value{
a list giving (i) the total number of genes observed and (ii) the number selected
}
\description{
Counts the total number of observed and selected entries across
a list of selection objects. The aim is to give the size of the complement of
the Venn diagramm.
}
\examples{

data(P2)
Q2=P2[1:3]

demo=function(x,a=500,b=1000){
  x$type="split"
  x$data=0*x$data
  x$data[a:b]=x$data[a:b]+1
  x
}

Q2=list(demo(P2[[1]],100,500),
        demo(P2[[2]],100,1000),
        demo(P2[[3]],5000,6000))

countUniverse(Q2)
}
