% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2tabs.r
\name{dataframe2tabs}
\alias{dataframe2tabs}
\title{converts a dataframe to a tab-deliminated charactor string}
\usage{
dataframe2tabs(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a charactor string
}
\description{
Converts a dataframe in to a character string which prints as a tab-deliminated file. Columns are
separated by tabs, rows by new-lines. The last character is a new line.
}
\examples{
dataframe2tabs(data.frame(a=c("one","two","three"),b=1:3,c=11:13))
}
