% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devPanel.r
\name{devPanel}
\alias{devPanel}
\title{test panel for development only}
\usage{
devPanel(id, input, output, session = NULL, ui = T, cx = cx)
}
\arguments{
\item{cx}{}
}
\description{
test panel for development only
}
