% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionReportPanel.r
\name{directionReportPanel}
\alias{directionReportPanel}
\title{direction informmation panel}
\usage{
directionReportPanel(id, input = NULL, output = NULL, ui = T, info = NULL)
}
\arguments{
\item{id}{Shiny ID for this widget}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{ui}{if TRUE, the UI is returned, otherwise the server is implemented}

\item{info}{a reactive values list from \link{pcaPanel} containing the information to be displayed.}
}
\value{

}
\description{
Creates a special panel for the display of information about selected directions in the PCA space. Currently supports
a selected principal component or a composite direction taken from the projection onto two or more PCs. Displays information
about the selection and a text input area from which weightings (loadings) can be downloaded. Information is displayed for
the last set of directional information saved.
}
