% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.r
\name{dummy}
\alias{dummy}
\title{dummy function}
\usage{
dummy(...)
}
\arguments{
\item{...}{any number of arguments, possibly zero}
}
\description{
Accepts an arbitary number of arguments but returns a NULL printing details of its call to the screen
}
\examples{
dummy()
dummy(1,2,3,4)
}
