% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoNames.r
\name{echoNames}
\alias{echoNames}
\title{sets the names of a character vector to have the same values as the vector itself}
\usage{
echoNames(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector but with names equal to itself
}
\description{
sets the names of a character vector to have the same values as the vector itself
}
\examples{
a=letters[1:5]
a

echoNames(a)
}
