% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.r
\name{fcat}
\alias{fcat}
\title{formated cat}
\usage{
fcat(format, ...)
}
\arguments{
\item{format}{a string format}

\item{...}{parameters passed after the format (e.g. values)}
}
\value{
a copy of the sctring printed to output
}
\description{
A version of \link[base]{sprintf} enclosed by a cat to write directly to the screen
}
\examples{
fcat("\%s = \%f","x",1.2)
fcat("\%s = (\%d,\%d)","a",34,54)
}
