% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenBundles.r
\name{flattenBundles}
\alias{flattenBundles}
\title{flatterns a list of bundles of source files into a list of source files}
\usage{
flattenBundles(bundles, usePrefix = T, annotateSource = T)
}
\arguments{
\item{bundles}{a list - ideally named - of bundles to be flatterned}

\item{usePrefix}{if true, the names of the input list are used to prefix the data sources, if false, the bundles text tag, bundle
is used instead. This is typically a text description of the experiment}

\item{annotateSource}{add prefix information to the descripton of each source file data type}
}
\value{
a list of source files
}
\description{
Projects and experiments are organised into bundles of grouped source files using the \link{goViewerInputSourceBundle}
function which can be selected for further analysis by the data source selection tab. Since the use can in principle
select multiple of source bundles, this are first selected as a list of bundles which then has to be flattenend into a normal
source file list used in analysis. This function supports this.
}
\examples{

# make some arbitary bundles

data(P2.source)

bundles=list(P2=goViewerInputSourceBundle(P2.source,
                                          "project two"),
             P2a=goViewerInputSourceBundle(P2.source[1:3],
                                           "test: P2 1-3"),
             P2b=goViewerInputSourceBundle(P2.source[4:5],
                                           "test: P2 4-5")
)

flattenBundles(bundles,usePrefix = T)
flattenBundles(bundles,usePrefix = F)
}
