% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatVennDiagram.r
\name{formatVennDiagram}
\alias{formatVennDiagram}
\title{formats a Venn diagram table for nicer display}
\usage{
formatVennDiagram(x, checked = "x")
}
\arguments{
\item{x}{a Venn diagram table generate by \link{computeVennDiagram}}

\item{checked}{symbol for included set, by default "x"}
}
\value{
a data frame showing an improved representation of x
}
\description{
A pretty formatter for \link{computeVennDiagram}. You may need to suppress any
quotes for printing.
}
\examples{
#' Venn=computeVennDiagram(list(c(4,1,2,3,5,6,7),
                          c(1,2,3,5,6,7,9),
                          c(5,6,7,9,8)))
Venn

formatVennDiagram(Venn)

# alternative symbols are supported but may not
# always display as expected on all systems
formatVennDiagram(Venn,"?")
}
