% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpaste.r
\name{fpaste}
\alias{fpaste}
\title{a formatted version of paste}
\usage{
fpaste(fmt, ...)
}
\arguments{
\item{fmt}{a \link{sprintf} type format, if a vector is concatonnated to a scalar with new
lines}

\item{...}{the parameters to be printed}
}
\value{
a character string
}
\description{
Use for printing reports to the screen etc.
}
\examples{
fpaste(c("today is the \%s",
         "I am \%s",
         "my score is \%d"),
         date(),"happy",8)

cat(fpaste(c("today is the \%s",
             "I am \%s",
             "my score is \%d"),
           date(),"happy",8))
}
