% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDownloadButton.r
\name{genericDownloadButton}
\alias{genericDownloadButton}
\title{Title}
\usage{
genericDownloadButton(id, input = NULL, output = NULL, ui = T, data = NULL)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Title
}
