% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalOptions.r
\name{globalOptions}
\alias{globalOptions}
\title{holder for general options}
\usage{
globalOptions(option = "")
}
\arguments{
\item{option}{name of the required option. If none, or an unidentified option is requested, a list
of all available options is returned.}
}
\value{
the value of the selected option
}
\description{
holder for general options
}
\examples{
globalOptions()
globalOptions("date")
globalOptions("colour.legend")
}
