% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.info.r
\name{GO.info}
\alias{GO.info}
\title{retrieves information for one or more GO terms}
\usage{
GO.info(go = c("GO:0006996", "GO:0006997"), terms = c("Term", "Ontology"))
}
\arguments{
\item{go}{a vector or one or more go terms for which information is to be displayed}

\item{terms}{a list of terms from the GO to be displayed}
}
\value{
a dataframe with rows for the requested ontologies and columns for the rows. The first column is the
   GO term ID.
}
\description{
retrieves information for one or more GO terms
}
\examples{
GO.info(c("GO:0042254","GO:0006996","GO:0006364"),
        c("Ontology","Term"))

# you can also get information by hand for any single GO

library(GO.db)
GOTERM[["GO:0006364"]]

## and a list of available slots (arguments for terms)

slotNames(GOTERM[["GO:0006364"]])
}
