% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysisDescripton.r
\name{goAnalysisDescripton}
\alias{goAnalysisDescripton}
\title{returns a descriptive text for a GO analysis}
\usage{
goAnalysisDescripton(contrast, GOtype, GOmethod, minNodeSize, ordered,
  noEnriched = NULL)
}
\arguments{
\item{noEnriched}{}
}
\description{
returns a descriptive text for a GO analysis
}
\examples{
goAnalysisDescripton("A vs B",
                     "BP",
                     "fisher",
                     100)

goAnalysisDescripton("A vs B",
                     "BP",
                     c("fisher","KS"),
                     100)

goAnalysisDescripton("A vs B",
                     "BP",
                     c("fisher","KS"),
                     100,500)
}
