% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goQQviewer.r
\name{goQQviewer}
\alias{goQQviewer}
\title{produce a observed vs expected quantile plot for GO enrichment}
\usage{
goQQviewer(id, input, output, session = NULL, ui = T, goObject = NULL,
  goResultsObject = NULL, goMethods = NULL, goSavedObject = NULL,
  title = NULL, title.nchar = 80, goStatus = NULL)
}
\arguments{
\item{id}{a Shiny UI id}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{session}{a Shiny output object}

\item{ui}{true generates the UI, false implements the server}

\item{goObject}{a gene ontology object (assumed reactive)}

\item{goResultsObject}{a gene ontology results object (assumed reactive)}

\item{goMethods}{a vector of one or more GO enrichment methods}

\item{title}{an optional title}

\item{title.nchar}{the number of characters per line in the title}

\item{goStatus}{a reactive list. If present, goStatus$ready=FALSE will inactivate this panel and cause it to return a NULL (blank) widget.}
}
\description{
Produces a QQ plot for a GO enrichment p-values
}
