% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goResultsDownload.r
\name{goResultsDownload}
\alias{goResultsDownload}
\title{downloads a copy of the GO output}
\usage{
goResultsDownload(id, input = NULL, output = NULL, ui = T,
  goObject = NULL, goResultsObject = NULL,
  fileName = "GO downloaded \%s.txt", message = T, goStatus = NULL)
}
\arguments{
\item{id}{root for the Shiny widget ID}

\item{input}{Shiny input object}

\item{output}{SHiny output object}

\item{ui}{if T returns UI, otherwise server side support}

\item{goObject}{a GO object from TopGO (optionally a reactive object)}

\item{goStatus}{a reactive list. If present, goStatus$ready=FALSE, the download button is surpressed and replaced by a message
instructing the user to first carry out a calculation.}

\item{goReultsObject}{a GO results object from TopGO (optinally a reactive object)}

\item{nameFormat}{the name of the output file. May not contain the
character % other than fillowed by an s - %s - where it is expanded
to the current data using \[base]{sprintf}.}
}
\description{
Implements a download button to store GO analysis outputs. Objects can be reactive
or no-reactive being excuted in the later case to make the code more portable.
OUtput is currently restricted to tab-deliminated .txt
}
