% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goViewerInput.r
\name{goViewerInput}
\alias{goViewerInput}
\title{constructs an input data object}
\usage{
goViewerInput(x, name = "input vector", annotation = "SYMBOL",
  type = "ordered", na.rm = F)
}
\arguments{
\item{x}{a named vector of gene classifications. This can either be binary to split the list into two groups or
numeric containing values which can be ordered for a later split by the tool. See type below. Gene names are
provided as the vectors names attribute using the convention defined by annotation. Alternatively, supplying a
DEseqResults object here will automatically populate x as an 'ordered' object}

\item{name}{a character string to be used as a name to identify the classfication}

\item{annotation}{a string defining the annotation type. Current the tool only supports SYMBOL, the default.}

\item{type}{either 'ordered' or 'split' corresponding to the numeric types described for x above.}

\item{na.rm}{if true, all removes missing values from x before returning}
}
\value{
an object of the class GoViewer.input
}
\description{
Constructs an input object for the GoViewer tool.
}
\examples{

genes=runif(10)
names(genes)=paste("gene",LETTERS[1:10])

goViewerInput(genes,
              name="test",
              annotation="test only")
}
