% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goViewerInputSource.r
\name{goViewerInputSource}
\alias{goViewerInputSource}
\title{constructs an input data source object}
\usage{
goViewerInputSource(
  x,
  name = "source data input",
  annotation = "SYMBOL",
  na.rm = F
)
}
\arguments{
\item{x}{a data frame or similar of gene or genomic data. Alternatively, supplying a
DEseqResults object here will automatically populate this object}

\item{name}{a character string to be used as a name to identify the classfication}

\item{annotation}{a string defining the annotation type. Current the tool only supports SYMBOL, the default.}

\item{na.rm}{if true, all removes missing values from x before returning}
}
\value{
an object of the class GoViewer.input
}
\description{
Constructs an input object for the GoViewer tool.
}
