% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goViewerInputSourceBundle.r
\name{goViewerInputSourceBundle}
\alias{goViewerInputSourceBundle}
\title{creates a bundle of input source objects}
\usage{
goViewerInputSourceBundle(x, bundle = "[no name given]")
}
\arguments{
\item{x}{a list of input source objects or a single input source object}

\item{bundle}{a name for the bundle, for example the study number or the experiment. This will be used
to identify it in the app.}
}
\value{
a GoViewer.source.bundle object
}
\description{
Creates a bundle of input source objects corresponding to say a study or an experiment. These will
be grouped together in the data selectton menu
}
\examples{
data(P2.source)
goViewerInputSourceBundle(P2.source,"project 2")
}
