% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goViewerSourceToInput.r
\name{goViewerSourceToInput}
\alias{goViewerSourceToInput}
\title{converts between a goViewer input source object and a simple imput object}
\usage{
goViewerSourceToInput(source, value = "p")
}
\arguments{
\item{source}{a input source object}

\item{value}{the data value upon which an ordering is to be based}
}
\value{
a goViewer input object
}
\description{
See \link{goViewerInput} and \link{goViewerInputSource} for details
}
\examples{
# a single case

data("P2.source")
names(P2.source)
x=P2.source[[1]]
x
head(x$data)

goViewerSourceToInput(x)

# or over a list

lapply(P2.source,goViewerSourceToInput)
}
