% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guard.r
\name{guard}
\alias{guard}
\title{guard function for unresolved Shiny input fields}
\usage{
guard(input, ns, names)
}
\arguments{
\item{input}{a shiny input object}

\item{ns}{a function generated by Shiny's \[shiny]{NS}}

\item{names}{a spaced deliminated string containing the required fields}
}
\value{
a logical scalar
}
\description{
Calling reactive functions before all required input fields have been resolved can cause warning messages
in the log and the screen. This function checks to see if all the required inputs are defined. If so,
a value of true is returned other-wise false. The names of the required inputs are given as a space
deliminated string and converted to Shiny input keys by a user supplied function ns generated by Shiny's
function \[Shiny]{NS}. This function is inspired on guard in Apple's Swift language.
}
\examples{
# simulate a Shiny function scenario
ns=NS("demo")
input=list("demo-a"=1:3,
           "demo-b"=NULL,
           "demo-c"=3:5)

guard(input,ns,"a b c")
guard(input,ns,"a c")
}
