% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvTab2.r
\name{gvTab2}
\alias{gvTab2}
\title{Title}
\usage{
gvTab2(id, input = NULL, output = NULL, session = NULL, ui = T,
  cx = NULL, helpLink = "https://www.bbc.co.uk/weather/")
}
\arguments{
\item{cx}{}
}
\description{
Title
}
