% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvTabDemo2.r
\name{gvTabDemo2}
\alias{gvTabDemo2}
\title{demo program for single tab installation}
\usage{
gvTabDemo2()
}
\value{

}
\description{
A simple code to illustrate the use of gvTab(). Three short sections of
code are required: A, B and C. See code listing for details.
}
\examples{
gvTabDemo()
}
