% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpPage.r
\name{helpPage}
\alias{helpPage}
\title{displays help panel}
\usage{
helpPage(
  id = NULL,
  input = NULL,
  output = NULL,
  ui = T,
  topic = NULL,
  links = "https://www.targetvalidation.org/"
)
}
\arguments{
\item{id}{a shiny ID (not currently used)}

\item{input}{a shiny input object (not currently used)}

\item{output}{a shiny output object (not currently used)}

\item{ui}{if true display the UI, o/w implement server (server side not currently used)}

\item{topic}{a code for the the topic to displayed}

\item{links}{one or more URLs to be displayed. Will be matchhed to %s strings in the template entry. Should be point to a URL
referecned by the document, Be default, points to the Open Targets website for testing purposes.}
}
\value{
output from a Shiny HTML tag
}
\description{

}
