% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputDataSummary.r
\name{inputDataSummary}
\alias{inputDataSummary}
\title{displays a summary of an input dataset}
\usage{
inputDataSummary(id, input, output, ui = T, x = NULL)
}
\arguments{
\item{id}{prefix for Shiny element IDs}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{ui}{true selects the UI, false the server}

\item{x}{a reactive function generating the selected input data}
}
\value{

}
\description{
Intended for development and debugging
}
