% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveLog10QqPlot.r
\name{interactiveLog10QqPlot}
\alias{interactiveLog10QqPlot}
\title{creates a simple -log10 p-value qqplot}
\usage{
interactiveLog10QqPlot(p, n = NA, thin = NA, print.top = F, ...)
}
\arguments{
\item{p}{a vector of p-values, possibly named}

\item{n}{if selected the number of smallest p-values to be annotated. Labels are taken from the names() of
p if provided: otherwise they are numbered in order of increasing p-value. The default NA labels no points.}

\item{thin}{if selected, thinks out all entries after the thin[1]-th element, reporting only the thin[2]-th. The thined
out elements are also printed in gray.}

\item{print.top}{prints the names of the top n genes if they are provided as names(p)}

\item{...}{optional additional parameters to be passed to \link[ggrepl]{geom_text_repel}.}
}
\value{
a list:
   \itemize{
      \item data - a copy of the data used to create the plot
      \item graph -ggplot2 plot object
      }
}
\description{
Constructs a simple log10 scale qq-plot for p-values. The log scale is preferred
to add focus to small highly significant points and since the angle made from
the origin is related to the corresponding FDR. The function also can be used to
annotate the top n p-values using the ggrepl package and thin out the plotting
of less signficant results.
}
