% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidated.r
\name{invalidated}
\alias{invalidated}
\title{tests if an expression is invalidated}
\usage{
invalidated(obj)
}
\arguments{
\item{obj}{a shiny reactive object}
}
\value{
a logical scalar
}
\description{
See https://stackoverflow.com/questions/45478521/listen-to-reactive-invalidation-in-shiny for details
}
