% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseNumberedList.r
\name{inverseNumberedList}
\alias{inverseNumberedList}
\title{returns a named list of sequential numbers}
\usage{
inverseNumberedList(x, list = T)
}
\arguments{
\item{x}{a vector of numbers or characters}

\item{list}{if false, a named vector is returned instead}
}
\value{
a named list or vector
}
\description{
Returns a list with elements 1 to the input's length labeled by a
character representation of the input.
Possible applications are selection lists in Shiny.
Optionally will also return a named vector.
}
\examples{
inverseNumberedList(c("Monday","Tuesday","Wednesday","Thursday","Friday"))
inverseNumberedList(c("Monday","Tuesday","Wednesday","Thursday","Friday"),list=F)
inverseNumberedList(1000:1010)
inverseNumberedList(1/(1:5))
}
