% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertList.r
\name{invertList}
\alias{invertList}
\title{swaps values for names in a named vector or flat atomic list}
\usage{
invertList(x)
}
\arguments{
\item{x}{a single depth list of atomic values. If nanes() is NULL, numbers are used}
}
\value{
an inverst copy of x with names and values transponsed
}
\description{
swaps values for names in a named vector or flat atomic list
}
\examples{
#'
x=list(a="alpha",b="beta",c="gamma")

x
invertList(x)

# but

y=list(a=1,b=2,c=3:4)
invertList(y)
}
