% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendPanel.r
\name{legendPanel}
\alias{legendPanel}
\title{generates a colour legend panel define colour legends}
\usage{
legendPanel(id, input, output, ui = T, cx = NULL)
}
\arguments{
\item{id}{Shiny widget id}

\item{input}{Shiny input object}

\item{output}{Shiny ouput object}

\item{ui}{generates UI if true, otherwise implements the server side}

\item{cx}{reactive list used here to read presets and assign back choices}
}
\description{
This panel communicates with the rest of the application via cx$colour.legends and cx$active.colour.legend.
}
