% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDataSources.r
\name{listDataSources}
\alias{listDataSources}
\title{lists available data}
\usage{
listDataSources(id, input, output, ui = T, cx = NULL)
}
\arguments{
\item{id}{widget id}

\item{input}{shiny input list}

\item{output}{shiney output list}

\item{ui}{true for ui, false for server}

\item{cx}{a reactive list containing data}
}
\description{
lists available data
}
