% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSelectionsPanel.r
\name{listSelectionsPanel}
\alias{listSelectionsPanel}
\title{list selected contrasts}
\usage{
listSelectionsPanel(id, input, output, ui, cx)
}
\arguments{
\item{id}{a Shiny UI id}

\item{input}{Shiny input}

\item{output}{Shiny output}

\item{ui}{selects UI code if true, o/w the server}

\item{cx}{a reactiveVector (c.f. pointer) to communicate data - here read the constrsts - with the remainder of the code}
}
\description{
Intended for development, this lists the contrasts selected by the contrast viewer
}
