% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainGoAnalysisPanel.r
\name{mainGoAnalysisPanel}
\alias{mainGoAnalysisPanel}
\title{main analysis panel}
\usage{
mainGoAnalysisPanel(id, input, output, session = NULL, ui = T,
  x.in = NULL, cx = NULL)
}
\arguments{
\item{x.in}{a GoViewerInput object requesting the current contrast}
}
\description{
main analysis panel
}
