% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGOobject.r
\name{makeGOobject}
\alias{makeGOobject}
\title{makes a basic GO object against which enrichment tests can be run}
\usage{
makeGOobject(x, goType = "BP", minNodeSize = 20, nPositives = 100, ...)
}
\arguments{
\item{x}{a data object as created by goViewerInput}

\item{goType}{character, type of gene ontology -}

\item{minNodeSize}{the minimum node size}

\item{nPositives}{the number of positive terms to be selected}
}
\value{

}
\description{
makes a basic GO object against which enrichment tests can be run
}
\examples{

# using our example dataset
data(P2)

P2a=P2[[1]]
P2a

head(P2a$data)

goObject=makeGOobject(P2a)
results=runTest(goObject,algorithm="classic",statistic = "fisher")

# displays results to screen
showSigOfNodes(goObject,score(results),firstSigNodes = 6,useInfo = 'all')
}
