% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeW.r
\name{makeW}
\alias{makeW}
\title{creates test projections from all possible views}
\usage{
makeW(z, sigma = 0.01)
}
\arguments{
\item{z}{a vector of any length}

\item{sigma}{the SD of the noise to be added to get observation}
}
\value{
a data frame. If z has length, this is a data frame with n rows and n(n-1)/2 columns. Each column
   contains two non-missing entries: the respective observations.
}
\description{
For a given vector, returns approximate views from all possible 2D axis combination views. Any required views can then be selected
by standard column subsetting. Non-observed co-ordinates are returned as NAs.
This routine is provided for testing and demonstration purposes.
}
\examples{
makeW(1:5)
}
