% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2p.r
\name{matrix2p}
\alias{matrix2p}
\title{computes a matrix of all length-p binary vectors}
\usage{
matrix2p(p)
}
\arguments{
\item{p}{integer, the vector length}
}
\value{
2^p x p  matrix
}
\description{
computes a matrix of all length-p binary vectors
}
\examples{
M=matrix2p(4)
}
