% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroredList.r
\name{mirroredList}
\alias{mirroredList}
\title{returns a list with names equal to content}
\usage{
mirroredList(x, list = T)
}
\arguments{
\item{x}{a vector}

\item{list}{if false, a named vector is returned}
}
\value{
a named list
}
\description{
returns a list with names equal to content
}
\examples{
mirroredList(1:3)
mirroredList(c("Monday","Tuesday","Wednesday"))
mirroredList(c("Monday","Tuesday","Wednesday"),list=F)

x=1:5
names(x)=LETTERS[1:5]
x
mirroredList(x,list=F)
}
