% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseL2.r
\name{normaliseL2}
\alias{normaliseL2}
\title{normalises a vector by its L2-norm}
\usage{
normaliseL2(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a normalised vector
}
\description{
Missing values are ignored.
}
\examples{
normaliseL2(c(1,1))
normaliseL2(c(1,1,1))
normaliseL2(c(1,1,NA))
}
