% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericColnames.r
\name{numericColnames}
\alias{numericColnames}
\title{returns the names of numeric dataframes columns}
\usage{
numericColnames(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
the names of numeric columns
}
\description{
Works like \link[base]{colnames} but only returns those which are numeric
}
\examples{
x=1:10
z=data.frame(a=x,
             b=as.factor(x),
             c=as.character(x),
             d=as.integer(x),
             e=as.Date(x,origin="2000-01-01"),
             f=x^2)
str(z)

numericColnames(z)
}
