% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2.r, R/data.r
\docType{data}
\name{P2}
\alias{P2}
\alias{P2}
\title{A list of RNA-seq datasets in GoViewer.input format for the following P2 comparisons, all without LPS stimulation}
\format{a list of GoViewer objects

\itemize{
  \item r1 - the baseline vs. PMA comparison
  \item r2 - the baseline vs. VD3 comparison
  \item r3 the THP-1 vs U937 comparison
  \item r1x  the change in r1 between THP-1 and U937
  \item r2x he change in r2 between THP-1 and U937
}}
\usage{
data(P2)

P2
}
\description{
A list of RNA-seq datasets in GoViewer.input format for the following P2 comparisons, all without LPS stimulation

A list of RNA-seq datasets in GoViewer.input format for the following P2 comparisons, all without LPS stimulation
}
\keyword{datasets}
