% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2.source.r
\docType{data}
\name{P2.source}
\alias{P2.source}
\title{A list of RNA-seq data views suitable for navigation and derivation of GoViewer input objects}
\format{
a list of GoViewer objects

\itemize{
  \item d1 - the baseline vs. PMA comparison
  \item d2 - the baseline vs. VD3 comparison
  \item d3 the THP-1 vs U937 comparison
  \item d1x  the change in r1 between THP-1 and U937
  \item d2x he change in r2 between THP-1 and U937
}
}
\usage{
data(P2.source)
}
\description{
A list of RNA-seq data views suitable for navigation and derivation of GoViewer input objects
}
\keyword{datasets}
