% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPanel.r
\name{pcaPanel}
\alias{pcaPanel}
\title{the PCA panel}
\usage{
pcaPanel(id, input, output, session = NULL, ui = T, cx = NULL)
}
\arguments{
\item{id}{shiny widget id}

\item{input}{shiny input}

\item{output}{shiny ouput}

\item{ui}{false selects server, true the UI}

\item{cx}{a reactive list for data}
}
\description{
the PCA panel
}
