% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLimmaVennDiagram.r
\name{plotLimmaVennDiagram}
\alias{plotLimmaVennDiagram}
\title{uses Lima to plot a graphically Venn diagram}
\usage{
plotLimmaVennDiagram(sets, allGenes = NULL)
}
\arguments{
\item{sets}{a list of vectors of gene names. The name of each list entry gives
gives the set name and the vector entries the set occupancy}

\item{allGenes}{a character vector. By default the total universe is simply the union of all of
sets. This option defines an alternative universe which may be larger or
smaller than the set of observed genes.}
}
\value{
produces a graph onto the screen
}
\description{
Makes a simple Venn diagram using \[limma]{vennDiagram} from the Limma package.
}
\examples{
plotLimmaVennDiagram(list(A=c("A","B"),
                          B=c("B","C","D"),
                          C=c("A","E")))

plotLimmaVennDiagram(list(A=c("A","B"),
                          B=c("B","C","D"),
                          C=c("A","E")),
                          allGenes=LETTERS)

plotLimmaVennDiagram(list(A=c("A","B"),
                          B=c("B","C","D"),
                          C=c("A","E")),
                     allGenes=c("A","B","C"))
}
