% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentPcaPointSelections.r
\name{presentPcaPointSelections}
\alias{presentPcaPointSelections}
\title{prepares output from readPcaPointSelections for display}
\usage{
presentPcaPointSelections(xyDir, additionalPoints, xyAutoReverse = F)
}
\arguments{
\item{xyDir}{a vector of length two giving the numbers of the principle components displayed. The first one should be for the
horizontal axis, the second for the vertical.}

\item{additionalPoints}{the $all output from \link{readPcaPointSelections}}

\item{xyAutoReverse}{reverses the x and y coordinates in xyDir[1] > xyDir[2]. This option is no longer required and defaults to FALSE}
}
\value{
a data frame of selected additional points with numerical rownames
}
\description{
Checks to see if any points have been identified in a given PCA view. If so, they are returned
in a dataset with sequential numerical rownames. Designed to show pairs of points used by the user
to specify directions on a PCA display for custom loadings.
}
\examples{
a=readPcaPointSelections(list(
  c(1,2,10,15),
  c(1,2,11,16),
  c(1,2,11,17),
  c(1,3,15,20),
  c(1,3,10,22),
  c(1,4,44,18)))

a

presentPcaPointSelections(1,2,a$all)
presentPcaPointSelections(1,3,a$all)
presentPcaPointSelections(1,4,a$all)
presentPcaPointSelections(1,5,a$all)
}
