% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GoViewer.input.r
\name{print.GoViewer.input}
\alias{print.GoViewer.input}
\title{prints an GoViewer data input object}
\usage{
print.GoViewer.input(x)
}
\arguments{
\item{x}{a GoView data input object}
}
\description{
prints an GoViewer data input object
}
\examples{
genes=runif(10)
names(genes)=paste("gene",LETTERS[1:10])

goViewerInput(genes,
              name="test",
              annotation="test only")
}
