% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.goViewerInputSourceBundle.r
\name{print.GoViewer.source.bundle}
\alias{print.GoViewer.source.bundle}
\title{prints a GoViewer.source.bundle}
\usage{
print.GoViewer.source.bundle(x)
}
\arguments{
\item{x}{a GoViewer.source.bundle object}
}
\description{
prints a GoViewer.source.bundle
}
\examples{

data(P2.source)
goViewerInputSourceBundle(P2.source,"project 2")
}
