% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GoViewer.source.r
\name{print.GoViewer.source}
\alias{print.GoViewer.source}
\title{prints an GoViewer data input source object}
\usage{
print.GoViewer.source(x)
}
\arguments{
\item{x}{a GoView data input object}
}
\description{
prints an GoViewer data input source object
}
