% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameListNames.r
\name{renameListNames}
\alias{renameListNames}
\title{assigns generic names to list slots}
\usage{
renameListNames(x)
}
\arguments{
\item{x}{a input list}
}
\value{
a copy of x with top-level slots renamed slot-1, slot-2, etc.
}
\description{
Forces a generic naming convention on a list of named or unnamed entries
}
\examples{
renameListNames(list(1:2,1:4))
}
