% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderValueToFront.r
\name{reorderValueToFront}
\alias{reorderValueToFront}
\title{moves a given value to the front of a vector}
\usage{
reorderValueToFront(x, value, order = F)
}
\arguments{
\item{x}{a vector of values}

\item{value}{the value or vector of values to be moved forward}

\item{order}{if true, returns the indices of the permuted vector labelled with the contents of the original array}
}
\value{
a vector of values potentially reordered
}
\description{
moves a given value to the front of a vector
}
\examples{
reorderValueToFront(letters[1:6],"d")
reorderValueToFront(letters[1:6],c("b","d"))

reorderValueToFront(letters[1:6],"z")
reorderValueToFront(letters[1:6],c("z","c"))

reorderValueToFront(letters[1:6],"d",order=T)
reorderValueToFront(letters[1:6],c("b","d"),order=T)
}
