% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerator.r
\name{reportGenerator}
\alias{reportGenerator}
\title{generate a downloadable report}
\usage{
reportGenerator(id, input = NULL, output = NULL, ui = T,
  buttonLabel = "download report", params = NULL,
  rmdScript = "report.rmd")
}
\arguments{
\item{id}{Shiny widget ID}

\item{input}{Shiny input opbject}

\item{output}{Shiny output opbject}

\item{ui}{if true, the UI output is created, otherwise server}

\item{buttonLabel}{a lable for the button (set from the SERVER side)}

\item{params}{an additional list of parameters to pass to the report generator}

\item{rmdScript}{the R markdown script to run. See \url{https://shiny.rstudio.com/articles/generating-reports.html} for a discussion}
}
\description{
This funcion provides a UI and serverside pair generate a written report from a .rmd markdown file. Currently only
PDF output is upported but additional formats could be provided later. Note that the UI side is called with the id argument set only:
the button itself (with any associated text) is created by the server. This ensures that the id's for the download action
are automatically matched. The report generation process can be slow, so a progress bar is provided. The markdown conversion appears broadly
stable but not yet 100%.
}
