% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestString.r
\name{requestString}
\alias{requestString}
\title{modal UI request for string}
\usage{
requestString(
  id.value,
  id.button,
  msg = "hello",
  msg.error = NULL,
  placeholder = "enter text here"
)
}
\arguments{
\item{id.value}{Shiny ID of the returned value}

\item{id.button}{shiny ID of the save button}

\item{msg}{prompting text}

\item{msg.error}{an addition message printed in red (typically an error message)}

\item{placeholder}{placeholder text}
}
\description{
modal UI request for string
}
