% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGOtests.r
\name{runGOtests}
\alias{runGOtests}
\title{runs one or more tests on a GO object}
\usage{
runGOtests(goObj, tests = "ks", algorithm = "classic", ...)
}
\arguments{
\item{goObj}{a defined GO generated by \link{makeGOobject}}

\item{tests}{a character vector of one or more tests. See \link[topGO]{whichTests} for details}

\item{algorithm}{a character string specifying the algorithm to be used. See \link[topGO]{whichWhich} for details}

\item{...}{additional options passed to runTest}
}
\value{
a list of GO results for one or more tests. Entries are named by tests.
}
\description{
A wrapper to pass multiple tests to \link[topGO]{runTest}.
}
\examples{
#basic example

P2a=P2[[1]]
P2a

head(P2a$data)

goObject=makeGOobject(P2a)
runObj=runGOtests(goObject,c("ks","fisher"))

GenTable(goObject,runObj[[1]])
GenTable(goObject,a=runObj[[1]],b=runObj[[2]])

# alternative front wrapper
tabulateGoResults(goObject,runObj)
}
