% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTool.r
\name{saveTool}
\alias{saveTool}
\title{widget to save and retrieve cx-list values}
\usage{
saveTool(id, input, output = NULL, cx = NULL, ui = T)
}
\arguments{
\item{id}{widget's Shiny ID}

\item{input}{input list}

\item{output}{output list}

\item{cx}{the cx list}

\item{ui}{TRUE (default) for UI calls, FALSE for server}
}
\description{
A development code to save and copy values of the cx reactive array. Doing this by hand
since it appears to be difficult to overwrite directly using load even if you get the same
target environments. Data is stored as a file cx.rda.
}
