% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityFilterPass.R
\name{securityFilterPass}
\alias{securityFilterPass}
\title{security check for selection formulae}
\usage{
securityFilterPass(
  x,
  allowedSymbols = c("x", "y", "e", "abs\\\\(", "\\\\)", "\\\\.", "=", "<", ">", "&",
    "\\\\|", "\\\\+", "-", "\\\\d")
)
}
\arguments{
\item{x}{string to be checked}

\item{allowedSymbols}{a list of allowed symbols. Symbols with a meaning in regular expressions should
be escaped as in the defaults.}
}
\value{
a logical: true if the string is OK, false if it contains unaccepted symbols.
}
\description{
A basic security check to allow simple expressions to specify regions to be safely evaluated from
user given input. Using the defaults, only expressions involving abs(), x and y, inequalities and
numbers are allowed. Unchecked, evaluated expressions can issue system's commands and hence
compromise a server.
}
\examples{

securityFilterPass("abs(x) > 2e+00 & y < 1e-10")
securityFilterPass("abs(x) > 2e+00 & y < 1e-10 | call(x)")
}
