% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBundles.r
\name{selectBundles}
\alias{selectBundles}
\title{selects group datasources (bundles) to be presented for analysis}
\usage{
selectBundles(id, input = NULL, output = NULL, ui = T, cx = NULL)
}
\arguments{
\item{id}{an id for the Shiny widget}

\item{input}{a Shiny input list}

\item{output}{a Shiny output list}

\item{cx}{a Shiny reactive list used to communicate data in and out}
}
\description{
Panel allows users to select bundles of data (e.g. experiments or projects) to include in their downstream analysis.
Any number of such bundles can be selected but as yet no choices within bundles is supported. The user can also decide
whether the full bundle description is carried forward as a prefix.
}
\details{
This function selects from a list cx$bundles and writes its selection to cx$contrast.data. By default (i.e. initially),
the first item cx$bundles[[1]] is selected
}
