% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputDataContrastView.r
\name{selectInputDataContrastView}
\alias{selectInputDataContrastView}
\title{selects the data for the contrast view}
\usage{
selectInputDataContrastView(D, x.var, x.metric, y.var, y.metric)
}
\arguments{
\item{D}{a data source object - see \link{goViewerInputsource} for details}

\item{x.var}{the response (e.g. a contrast) to be plotted on the x-axis}

\item{x.metric}{the metric of x.var to be plotted}

\item{y.var}{the response for the y-axis}

\item{y.metric}{the metric of the y-axis response}
}
\value{
a three column data frame with three columns: name, x and y where name is the
   gene or entity tested. If no match can be found, NULL is returned
}
\description{
Data is returned as three column dataset or NULL if no match can be found.
}
\examples{
data(P2.source)

z=selectInputDataContrastView(P2.source,
                      "d1","l2fc",
                      "d1","p")
head(z)
}
