% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputDataObject.r
\name{selectInputDataObject}
\alias{selectInputDataObject}
\title{selects the input contrast}
\usage{
selectInputDataObject(id, input, output, ui, all.inputs = NULL)
}
\arguments{
\item{all.inputs}{a reactive data object containing all the avaliable contrasts}
}
\value{

}
\description{
selects the input contrast
}
