% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setListHelper.r
\name{setListHelper}
\alias{setListHelper}
\title{helper function to extend binary set functions to lists of sets}
\usage{
setListHelper(z, op = intersect)
}
\arguments{
\item{z}{a list of sets, each represented as a vector}

\item{op}{a bnary function, default is intersect()}
}
\value{
a vector representation of a set
}
\description{
helper function to extend binary set functions to lists of sets
}
\examples{
setListHelper(list(1:10,2:4,1:3),intersect)
setListHelper(list(1:10,2:4,1:3),union)
}
