% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortArrow.r
\name{shortArrow}
\alias{shortArrow}
\title{a shorten arrow}
\usage{
shortArrow(x0, y0, x1, y1, alpha = 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Not used.Need something for ggplot2, but this is base graphics
}
\examples{
plot(c(1,2),c(2,4),xlim=c(0,3),ylim=c(1,5))
shortArrow(1,2,2,4,0.80,col="red")
}
