% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScoresPlot.r
\name{simpleScoresPlot}
\alias{simpleScoresPlot}
\title{a simple scores plot suitable for PCA or other decompositions}
\usage{
simpleScoresPlot(
  scores.data,
  scores.sd = NULL,
  pcs = 2:3,
  size = 2,
  label = "comp",
  rows = NULL,
  axes = T,
  reflect.x = F,
  reflect.y = F,
  scale.x = 1,
  scale.y = 1,
  colourDictionary = NULL,
  additionalPoints = NULL,
  additionalPointsWithArrows = T
)
}
\arguments{
\item{scores.data}{a scores matrix with columns for components.
Assumes items are labelled as rownames}

\item{scores.sd}{a vector of associated SD's in the case of PCA. If not present, NULL provides
a default output. Note that these are the \emph{square roots} if the eigenvalues}

\item{pcs}{a vector of the components to be compared. If length greater than two, are binary
combinations are run and printed to the screen}

\item{size}{size parameter for label texts}

\item{label}{the prefix used for axis labels if the scores.data columnames are not defined.
Defaults to "PC". a value of "PC" with a pcs of 1:2
would generate axis labels of PC1 and PC2.}

\item{rows}{a convenience argument to add rownames to the scores data array}

\item{axes}{default true, causes reference lines to be drawn for the horizontal
vertical axes}

\item{reflect.x}{if true, reverses the horizontal axis}

\item{reflect.y}{if true, reverses the vertical axis}

\item{scale.x}{optional parameter to scale the x axis data prior to plotting.
Default value is 1. Values of -1 may be useful to adjust signs to adjust
for other data terms. Now how this differs from reflect.x which reverses
the way the axis is constructed but does not break the relationship between
its values and that of the data.}

\item{scale.y}{as scale.x but for the y-axis. Default value 1.}

\item{colourDictionary}{an optional colour lookup table. See \link{legendColourLookUp} for details}

\item{additionalPoints}{an optional data frame of additional points to be plotted over the standard output. Positions are
read from the \link{assumed} x and y columns and the text to be plotted is taken from its rownames. If no row names are
supplied, points are labelled sequentially, 1, 2, 3 etc.}

\item{additionalPointsWithArrows}{if additional points selected, setting this to true
links them with arrows}
}
\value{
a ggplot object containing the required graph. If length(pcs) > 2, NULL is returned
but all indicated views are printed to the screen
}
\description{
Typically called by
}
