% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedAnalyses.r
\name{storedAnalyses}
\alias{storedAnalyses}
\title{reproduces stored GO analysis to the screen}
\usage{
storedAnalyses(
  id,
  input = NULL,
  output = NULL,
  session = NULL,
  ui = T,
  cx = NULL
)
}
\arguments{
\item{id}{a shiny widget ID}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{ui}{TRUE selects user interface components, o/w server}

\item{cx}{a shiny reactive list containing the stored analyses in $results}
}
\value{

}
\description{
Displays previous results stored in cx$results. If this list is empty, a no analysis
message is displayed. Currently only displays plots but will extend to include tables
}
