% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyBundle.r
\name{studyBundle}
\alias{studyBundle}
\title{creates}
\usage{
studyBundle(
  de = NULL,
  epigen = NULL,
  study = "study 2",
  description = "a part of the "
)
}
\arguments{
\item{de}{a list goViewerInputSource objects}

\item{epigen}{an object containing cross plaform expression and epigenetics data}

\item{study}{the studies name}

\item{description}{a study description}
}
\value{

}
\description{
creates
}
