% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulateGoResults.r
\name{tabulateGoResults}
\alias{tabulateGoResults}
\title{gives a tabular summary of the top k GOs}
\usage{
tabulateGoResults(goObject, runObj, k = 20, ...)
}
\arguments{
\item{goObject}{a defined GO generated by \link{makeGOobject}}

\item{runObj}{a list of GO tests results returned by \link{runGOtests}}

\item{k}{the number of GO's to be listed}

\item{...}{additional arguments passed ot \link[topGO]{GenTable}}
}
\value{
a table of the top k GOs returned by \link[topGO]{GenTable}
}
\description{
gives a tabular summary of the top k GOs
}
\examples{
#basic example

P2a=P2[[1]]
P2a

head(P2a$data)

goObject=makeGOobject(P2a)
runObj=runGOtests(goObject,c("ks","fisher"))

GenTable(goObject,runObj[[1]])
GenTable(goObject,a=runObj[[1]],b=runObj[[2]])

# alternative front wrapper
tabulateGoResults(goObject,runObj)
}
