% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textMessageAsPlot.r
\name{textMessageAsPlot}
\alias{textMessageAsPlot}
\title{plot message to screen as a plot}
\usage{
textMessageAsPlot(msg = "wait ...")
}
\arguments{
\item{msg}{the message to be printed. Multiline messages are supported}
}
\description{
Prints text to the screen as a base graphics plot. Note, no border is
printed.
}
\examples{
textMessageAsPlot("hello")
textMessageAsPlot(sprintf("hello world!\nthe current time\n is \%s\nsee you again\nsoon",date()))
}
