% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplainedSummary.r
\name{varianceExplainedSummary}
\alias{varianceExplainedSummary}
\title{prints a variance explained table from a prcomp PCA output}
\usage{
varianceExplainedSummary(pca.obj, k = NA)
}
\arguments{
\item{pca.obj}{a prcomp PCA object}

\item{k}{the number of PC's to be included. If NA, all PCs are included}
}
\value{
a data frame of percentages and cumulatived percentages explained
}
\description{
prints a variance explained table from a prcomp PCA output
}
\examples{
pca.obj=prcomp(USArrests)
summary(pca.obj)

varianceExplainedSummary(pca.obj)
varianceExplainedSummary(pca.obj,3)
}
