% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2string.r
\name{vector2string}
\alias{vector2string}
\title{prints a single vector as a string}
\usage{
vector2string(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a string
}
\description{
Prints a vector as a string with comma deliminators. See examples for details
}
\examples{
vector2string(1:4)
vector2string(signif(1/1:4,3))
vector2string(letters[1:5])
}
