% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagramDownload.r
\name{vennDiagramDownload}
\alias{vennDiagramDownload}
\title{creates a download page for the Venn diagram panel}
\usage{
vennDiagramDownload(id, input, output, ui = T, setSelections = NULL)
}
\arguments{
\item{id}{shiny ID}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{ui}{true if use in UI mode}

\item{setSelections}{a reactive object created by \link{vennDiagramPanel}}
}
\description{
creates a download page for the Venn diagram panel
}
