% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagramPanel.r
\name{vennDiagramPanel}
\alias{vennDiagramPanel}
\title{creates a tab view for Venn diagrams}
\usage{
vennDiagramPanel(id, input, output, ui = T, cx = NULL)
}
\arguments{
\item{cx}{}
}
\value{

}
\description{
creates a tab view for Venn diagrams
}
